Add --count option
The -c, --count command line option is included to allow the user to
see the number of elements in each JSON array for each top-level key in
a JSON object.

Since -c is now used for --count, -c for --nocolor has been changed to
-n.

Since jsoncut operates on sequences, i.e. JSON arrays this is useful
for determining the rootkey with the most elements in an array; for
example an API might retrieve a list of items which are stored in one
of it's keys.  If the user is not familiar with the data structure this
may assist them in quickly locating the key that contains those items.