Change default for --any flag option to True
It's common for keys to not always exist in every record in a JSON
document; most of the time the user will just want it dump out what
it can find and ignore any keys that are not present.  Previously it
would stop whenver it couldn't find a key that was in the key list.